Move agent proxy to main mediator package
