+++ mediator/src/main/java/cas/ypsiliform/mediator/MediatorFactory.java
@@ -2,7 +2,6 @@ package cas.ypsiliform.mediator;
 
 import java.util.Map;
 
-import cas.ypsiliform.mediator.negotiation.AgentProxy;
 import cas.ypsiliform.mediator.negotiation.VoteBasedMediator;
 
 public class MediatorFactory {
